:truck: Move SHM stack components into 'stacks'
