+++ data_safe_haven/infrastructure/stacks/shm/bastion.py
