+++ data_safe_haven/infrastructure/stacks/shm/data.py
