+++ data_safe_haven/infrastructure/stacks/shm/monitoring.py
