+++ data_safe_haven/infrastructure/stacks/shm/networking.py
